[Ubuntu] don't install web servers
Inside container environment we're not running a second service process,
but use service container for act instead.

See: https://github.com/nektos/act/pull/1949

+++ images/ubuntu/scripts/build/install-apache.sh
@@ -4,6 +4,10 @@
 ##  Desc:  Install Apache HTTP Server
 ################################################################################
 
+if [[ -f /run/systemd/container ]]; then
+    exit 0
+fi
+
 # Install Apache
 apt-get install apache2 -y
